Trait taple_core::ApiModuleInterface

pub trait ApiModuleInterface {

    // Required methods
    fn external_request<'life0, 'async_trait>(
        &'life0 self,
        event_request: Signed<EventRequest>
    ) -> Pin<Box<dyn Future<Output = Result<DigestIdentifier, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_subjects<'life0, 'async_trait>(
        &'life0 self,
        namespace: String,
        from: Option<String>,
        quantity: Option<i64>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_governances<'life0, 'async_trait>(
        &'life0 self,
        namespace: String,
        from: Option<String>,
        quantity: Option<i64>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_subjects_by_governance<'life0, 'async_trait>(
        &'life0 self,
        governance_id: DigestIdentifier,
        from: Option<String>,
        quantity: Option<i64>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_events<'life0, 'async_trait>(
        &'life0 self,
        subject_id: DigestIdentifier,
        from: Option<i64>,
        quantity: Option<i64>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Signed<Event>>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_event<'life0, 'async_trait>(
        &'life0 self,
        subject_id: DigestIdentifier,
        sn: u64
    ) -> Pin<Box<dyn Future<Output = Result<Signed<Event>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_subject<'life0, 'async_trait>(
        &'life0 self,
        subject_id: DigestIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<SubjectData, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn shutdown<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = Result<(), ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait;
    fn approval_request<'life0, 'async_trait>(
        &'life0 self,
        request_id: DigestIdentifier,
        acceptance: bool
    ) -> Pin<Box<dyn Future<Output = Result<ApprovalEntity, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_pending_requests<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ApprovalEntity>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_single_request<'life0, 'async_trait>(
        &'life0 self,
        id: DigestIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<ApprovalEntity, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn add_preauthorize_subject<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        subject_id: &'life1 DigestIdentifier,
        providers: &'life2 HashSet<KeyIdentifier>
    ) -> Pin<Box<dyn Future<Output = Result<(), ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_all_allowed_subjects_and_providers<'life0, 'async_trait>(
        &'life0 self,
        from: Option<String>,
        quantity: Option<i64>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<(DigestIdentifier, HashSet<KeyIdentifier>)>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn add_keys<'life0, 'async_trait>(
        &'life0 self,
        derivator: KeyDerivator
    ) -> Pin<Box<dyn Future<Output = Result<KeyIdentifier, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_validation_proof<'life0, 'async_trait>(
        &'life0 self,
        subject_id: DigestIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<(HashSet<Signature>, ValidationProof), ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_request<'life0, 'async_trait>(
        &'life0 self,
        request_id: DigestIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<TapleRequest, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_governance_subjects<'life0, 'async_trait>(
        &'life0 self,
        governance_id: DigestIdentifier,
        from: Option<String>,
        quantity: Option<i64>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_approval<'life0, 'async_trait>(
        &'life0 self,
        request_id: DigestIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<ApprovalEntity, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_approvals<'life0, 'async_trait>(
        &'life0 self,
        state: Option<ApprovalState>,
        from: Option<String>,
        quantity: Option<i64>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ApprovalEntity>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Trait that allows implementing the interface of a TAPLE node. The only native implementation is NodeAPI. Users can use the trait to add specific behaviors to an existing node interface. For example, a NodeAPI wrapper could be created that again implements the trait and perform certain intermediate operations, such as incrementing a counter to find out how many API queries have been made.

Required Methods

fn external_request<'life0, 'async_trait>( &'life0 self, event_request: Signed<EventRequest> ) -> Pin<Box<dyn Future<Output = Result<DigestIdentifier, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Allows to make a request to the node from an external Invoker

fn get_subjects<'life0, 'async_trait>( &'life0 self, namespace: String, from: Option<String>, quantity: Option<i64> ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Allows to get all subjects that are known to the current node, regardless of their governance. Paging can be performed using the optional arguments from and quantity. Regarding the first one, note that it admits negative values, in which case the paging is performed in the opposite direction starting from the end of the collection. Note that this method also returns the subjects that model governance.

Possible errors

• [ApiError::InternalError] if an internal error occurred during the execution of the operation.

fn get_governances<'life0, 'async_trait>( &'life0 self, namespace: String, from: Option<String>, quantity: Option<i64> ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

It allows to obtain all the subjects that model existing governance in the node.

Possible errors

• [ApiError::InternalError] if an internal error occurred during the execution of the operation.

fn get_subjects_by_governance<'life0, 'async_trait>( &'life0 self, governance_id: DigestIdentifier, from: Option<String>, quantity: Option<i64> ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn get_events<'life0, 'async_trait>( &'life0 self, subject_id: DigestIdentifier, from: Option<i64>, quantity: Option<i64> ) -> Pin<Box<dyn Future<Output = Result<Vec<Signed<Event>>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Allows to obtain events from a specific subject previously existing in the node. Paging can be performed by means of the optional arguments from and quantity. Regarding the former, it should be noted that negative values are allowed, in which case the paging is performed in the opposite direction starting from the end of the string.

Possible errors

• ApiError::InvalidParameters if the specified subject identifier does not match a valid DigestIdentifier.

fn get_event<'life0, 'async_trait>( &'life0 self, subject_id: DigestIdentifier, sn: u64 ) -> Pin<Box<dyn Future<Output = Result<Signed<Event>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn get_subject<'life0, 'async_trait>( &'life0 self, subject_id: DigestIdentifier ) -> Pin<Box<dyn Future<Output = Result<SubjectData, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Allows to obtain a specified subject by specifying its identifier.

Possible errors

• ApiError::InvalidParameters if the specified identifier does not match a valid DigestIdentifier.
• ApiError::NotFound if the subject does not exist.

fn shutdown<'async_trait>( self ) -> Pin<Box<dyn Future<Output = Result<(), ApiError>> + Send + 'async_trait>>where Self: 'async_trait,

Stops the node, consuming the instance on the fly. This implies that any previously created API or [NotificationHandler] instances will no longer be functional.

fn approval_request<'life0, 'async_trait>( &'life0 self, request_id: DigestIdentifier, acceptance: bool ) -> Pin<Box<dyn Future<Output = Result<ApprovalEntity, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Allows to vote on a voting request that previously exists in the system. This vote will be sent to the corresponding node in charge of its collection.

Possible errors

• [ApiError::InternalError] if an internal error occurs during operation execution.
• ApiError::NotFound if the request does not exist in the system.
• ApiError::InvalidParameters if the specified request identifier does not match a valid DigestIdentifier.
• ApiError::VoteNotNeeded if the node’s vote is no longer required.
This occurs when the acceptance of the changes proposed by the petition has already been resolved by the rest of the nodes in the network or when the node cannot participate in the voting process because it lacks the voting role.

fn get_pending_requests<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<ApprovalEntity>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

It allows to obtain all the voting requests pending to be resolved in the node. These requests are received from other nodes in the network when they try to update a governance subject. It is necessary to vote their agreement or disagreement with the proposed changes in order for the events to be implemented.

Possible errors

• [ApiError::InternalError] if an internal error occurs during operation execution.

fn get_single_request<'life0, 'async_trait>( &'life0 self, id: DigestIdentifier ) -> Pin<Box<dyn Future<Output = Result<ApprovalEntity, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

It allows to obtain a single voting request pending to be resolved in the node. This request is received from other nodes in the network when they try to update a governance subject. It is necessary to vote its agreement or disagreement with the proposed changes in order for the events to be implemented.

Possible errors

• [ApiError::InternalError] if an internal error occurs during operation execution. • ApiError::NotFound if the requested request does not exist.

fn add_preauthorize_subject<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, subject_id: &'life1 DigestIdentifier, providers: &'life2 HashSet<KeyIdentifier> ) -> Pin<Box<dyn Future<Output = Result<(), ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

fn get_all_allowed_subjects_and_providers<'life0, 'async_trait>( &'life0 self, from: Option<String>, quantity: Option<i64> ) -> Pin<Box<dyn Future<Output = Result<Vec<(DigestIdentifier, HashSet<KeyIdentifier>)>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn add_keys<'life0, 'async_trait>( &'life0 self, derivator: KeyDerivator ) -> Pin<Box<dyn Future<Output = Result<KeyIdentifier, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn get_validation_proof<'life0, 'async_trait>( &'life0 self, subject_id: DigestIdentifier ) -> Pin<Box<dyn Future<Output = Result<(HashSet<Signature>, ValidationProof), ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn get_request<'life0, 'async_trait>( &'life0 self, request_id: DigestIdentifier ) -> Pin<Box<dyn Future<Output = Result<TapleRequest, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn get_governance_subjects<'life0, 'async_trait>( &'life0 self, governance_id: DigestIdentifier, from: Option<String>, quantity: Option<i64> ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn get_approval<'life0, 'async_trait>( &'life0 self, request_id: DigestIdentifier ) -> Pin<Box<dyn Future<Output = Result<ApprovalEntity, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn get_approvals<'life0, 'async_trait>( &'life0 self, state: Option<ApprovalState>, from: Option<String>, quantity: Option<i64> ) -> Pin<Box<dyn Future<Output = Result<Vec<ApprovalEntity>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,