Crate taple_core

TAPLE is a DLT focused on traceability characterized by its level of scalability, its flexibility to be employed in different devices and use cases and its reduced resource consumption, including power consumption.

The TAPLE crate provides the library that allows instantiating nodes of this DLT in order to create a functional network through a single structure containing all the required logic. Applications can interact with these nodes through the API they expose, thus enabling read and write operations against the network. The API also allows the design and creation of customized clients for the technology according to the user’s needs.

In addition to the node itself, the library also exposes a series of data structures specific to the protocol that can be obtained when interacting with the API or, in some cases, may be necessary to interact with it.

Basic usage

use std::str::FromStr;
use std::time::Duration;

use taple_core::ListenAddr;
use taple_core::Notification;
use taple_core::crypto;
use taple_core::crypto::Ed25519KeyPair;
use taple_core::crypto::KeyGenerator;
use taple_core::crypto::KeyMaterial;
use taple_core::get_default_settings;
use taple_core::request::StartRequest;
use taple_core::signature::Signature;
use taple_core::signature::Signed;
use taple_core::ApiModuleInterface;
use taple_core::Derivable;
use taple_core::DigestIdentifier;
use taple_core::EventRequest;
use taple_core::KeyDerivator;
use taple_core::KeyIdentifier;
use taple_core::MemoryManager;
use taple_core::Taple;

#[tokio::main]
async fn main() -> Result<(), ()> {
   let tmp_path = std::env::temp_dir();
   let path = format!("{}/.taple/contracts", tmp_path.display());
   std::fs::create_dir_all(&path).expect("Contract folder not created");

   let mut settings = get_default_settings();
   settings.network.listen_addr = vec![ListenAddr::Memory { port: Some(40040) }];
   settings.node.smartcontracts_directory = path;
   // Generate ramdon cryptographic material
   let keypair = Ed25519KeyPair::from_seed(&[0; 32]);
   let hex_private_key = hex::encode(&keypair.secret_key_bytes());
   settings.node.secret_key = Some(hex_private_key);
   let kp = crypto::KeyPair::Ed25519(keypair);
   let public_key = kp.public_key_bytes();
   let key_identifier = KeyIdentifier::new(kp.get_key_derivator(), &public_key);

   // We call the constructor method of a TAPLE node
   let mut taple = Taple::new(settings, MemoryManager::new());
   // We can use the next structure to get notifications from the node
   let mut notifier = taple.get_notification_handler();
   // The TAPLE node generates several Tokyo tasks to manage the different
   // components of its architecture.
   // The "start" method initiates these tasks and returns the control flow.
   let shutdown_manager = taple.get_shutdown_manager();
   taple.start().await.expect("TAPLE started");
   // From this point the user can start interacting with the node.
   // It is the user's responsibility to decide whether to keep the node running.
   // To do so, the main thread of the application must not terminate.
   let api = taple.get_api();

   // First we need to add the keys for our Subject
   let pk_added = api
       .add_keys(KeyDerivator::Ed25519)
       .await
       .expect("Error getting server response");

   // Create the request
   let event_request = EventRequest::Create(StartRequest {
       governance_id: DigestIdentifier::default(),
       name: "Paco23".to_string(),
       namespace: "".to_string(),
       schema_id: "governance".to_string(),
       public_key: pk_added,
   });

   let signature = Signature::new(&event_request, key_identifier, &kp).unwrap();

   let er_signed = Signed::<EventRequest> {
       content: event_request.clone(),
       signature,
   };

   // We need to create the governance, the game set of rules of our future network, to start creating subject on it.
   let _request_id = api.external_request(er_signed).await.unwrap();
   // We wait until the notification of a new subject created is received
   let subject_id = tokio::select! {
       _ = tokio::time::sleep(Duration::from_millis(5000)) => {
           panic!("Max time waited");
       },
       subject_id = async {
           loop {
               match notifier.receive().await {
                   Ok(data) => {
                       if let Notification::NewSubject { subject_id } = data {
                           break subject_id;
                       }
                   },
                   Err(_) => {
                       panic!("Notifier error")
                   }
               }
           }
       } => subject_id
   };

   // Next, we convert the subject_id to a DigestIdentifier
   let subject_id = DigestIdentifier::from_str(&subject_id).expect("DigestIdentifier from str failed");

   // We use this ID to retrieve governance data
   let subject = api.get_subject(subject_id.clone()).await.expect(&format!(
       "Error getting subject content with id: {}",
       subject_id.to_str()
   ));

   println!("Governance subject Id: {:#?}", subject.subject_id.to_str());
   println!("Governance subject SN: {:#?}", subject.sn);

   // Now we send a signal to stop our TAPLE node:
   shutdown_manager.shutdown().await;
   Ok(())
}

Re-exports

pub use error::Error;

Modules

crypto

This module provides the structs and traits for the generation of key pairs for cryptographic operations.

error

Possible errors of a TAPLE Node

event

identifier

Identifiers module

protocol

request

signature

Define the data structures related to signatures

Macros

test_database_manager_trait

Structs

ApprovalEntity

A struct representing an approval entity.

ApprovalRequest

A struct representing an approval request.

ApprovalResponse

A struct representing an approval response.

DigestIdentifier

Digest based identifier

EvaluationRequest

A struct representing an evaluation request.

EvaluationResponse

A struct representing an evaluation response.

Event

A struct representing an event.

GetEvents

GetSubject

GetSubjects

KeyIdentifier

Key based identifier

MemoryManager

Metadata

A struct representing the metadata of a TAPLE event.

NetworkSettings

P2P network configuration parameters of a TAPLE node.

NodeAPI

Object that allows interaction with a TAPLE node.

NodeSettings

General settings of a TAPLE node.

NotificationHandler

Object that allows receiving notifications of the different events of relevance that a node performs and/or detects.

SignatureIdentifier

Signature based identifier

SubjectData

A struct representing the data associated with a TAPLE subject.

Taple

Structure representing a node of a TAPLE network.

TapleSettings

Configuration parameters of a TAPLE node divided into categories.

TapleShutdownManager

TimeStamp

A struct representing a timestamp.

ValidationProof

A struct representing a validation proof.

ValueWrapper

A struct representing a wrapper around a JSON value.

Enums

APICommands

ApiError

Errors that may occur when using the TAPLE API

ApiResponses

ApprovalState

An enum representing the state of an approval entity.

DbError

DigestDerivator

Enumeration with digest derivator types

EventRequest

An enum representing a TAPLE event request.

KeyDerivator

An enumeration of key derivator types.

ListenAddr

ListenAddrErrors

Notification

Notifications generated by [NotificationHandler]. These notifications identify the type of message and its content. In addition, the message is accompanied by the data used to construct the message, so that it can be used to construct customized messages.

Traits

ApiModuleInterface

Trait that allows implementing the interface of a TAPLE node. The only native implementation is NodeAPI. Users can use the trait to add specific behaviors to an existing node interface. For example, a NodeAPI wrapper could be created that again implements the trait and perform certain intermediate operations, such as incrementing a counter to find out how many API queries have been made.

DatabaseCollection

A trait representing a collection of key-value pairs in a database.

DatabaseManager

Derivable

Derivable Identifiers

Functions

get_default_settings

Instance a default settings to start a new Taple Node