Module taple_core::crypto

This module provides the structs and traits for the generation of key pairs for cryptographic operations.

Structs

BaseKeyPair

Base for asymmetric key pair

CryptoBox

Represents asymetric key pair for storage (deprecated: KeyPair is serializable)

Enums

KeyPair

Asymmetric key pair

Payload

Payloads

Traits

DHKE

Used for Diffie–Hellman key exchange operations

DSA

Used for Digital Signature Algorithm (DSA)

KeyGenerator

Collection of methods to initialize a key pair using random or deterministic manner

KeyMaterial

Return key material bytes

Functions

create_seed

Creates 32 bytes seed

generate

Generate key pair

Type Definitions

Ed25519KeyPair

Ed25519 cryptographic key pair

Secp256k1KeyPair

Secp256k1 cryptographic key pair