Trait taple_core::ApiModuleInterface

pub trait ApiModuleInterface {

    // Required methods
    fn create_request<'life0, 'async_trait>(
        &'life0 self,
        request_type: EventRequestType
    ) -> Pin<Box<dyn Future<Output = Result<DigestIdentifier, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn external_request<'life0, 'async_trait>(
        &'life0 self,
        event_request: EventRequest
    ) -> Pin<Box<dyn Future<Output = Result<DigestIdentifier, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_event<'life0, 'async_trait>(
        &'life0 self,
        subject_id: DigestIdentifier,
        payload: String
    ) -> Pin<Box<dyn Future<Output = Result<DigestIdentifier, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_all_subjects<'life0, 'async_trait>(
        &'life0 self,
        namespace: String,
        from: Option<String>,
        quantity: Option<i64>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_all_governances<'life0, 'async_trait>(
        &'life0 self,
        namespace: String,
        from: Option<String>,
        quantity: Option<i64>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_event_of_subject<'life0, 'async_trait>(
        &'life0 self,
        subject_id: DigestIdentifier,
        from: Option<i64>,
        quantity: Option<i64>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Event>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_subject<'life0, 'async_trait>(
        &'life0 self,
        governance_id: DigestIdentifier,
        schema_id: String,
        namespace: String
    ) -> Pin<Box<dyn Future<Output = Result<DigestIdentifier, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_subject<'life0, 'async_trait>(
        &'life0 self,
        subject_id: DigestIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<SubjectData, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_governance<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<DigestIdentifier, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn shutdown<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = Result<(), ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait;
    fn add_preauthorize_subject<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        subject_id: &'life1 DigestIdentifier,
        providers: &'life2 HashSet<KeyIdentifier>
    ) -> Pin<Box<dyn Future<Output = Result<(), ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn expecting_transfer<'life0, 'async_trait>(
        &'life0 self,
        subject_id: DigestIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<KeyIdentifier, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_validation_proof<'life0, 'async_trait>(
        &'life0 self,
        subject_id: DigestIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<HashSet<Signature>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_governance_subjects<'life0, 'async_trait>(
        &'life0 self,
        governance_id: DigestIdentifier,
        from: Option<String>,
        quantity: Option<i64>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Trait that allows implementing the interface of a TAPLE node. The only native implementation is NodeAPI. Users can use the trait to add specific behaviors to an existing node interface. For example, a NodeAPI wrapper could be created that again implements the trait and perform certain intermediate operations, such as incrementing a counter to find out how many API queries have been made.

Required Methods

fn create_request<'life0, 'async_trait>( &'life0 self, request_type: EventRequestType ) -> Pin<Box<dyn Future<Output = Result<DigestIdentifier, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Allows to generate a voting request in the system. This request will be sent to the node that will be in charge of handling the votes of the rest of the nodes belonging to the same governance.

Possible errors

• [ApiError::InternalError] if an internal error occurs during operation execution.
• ApiError::InvalidParameters if the specified request identifier does not match a valid DigestIdentifier.

fn external_request<'life0, 'async_trait>( &'life0 self, event_request: EventRequest ) -> Pin<Box<dyn Future<Output = Result<DigestIdentifier, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Allows to make a request to the node from an external Invoker

fn create_event<'life0, 'async_trait>( &'life0 self, subject_id: DigestIdentifier, payload: String ) -> Pin<Box<dyn Future<Output = Result<DigestIdentifier, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Allows adding a new event to the chain of a subject previously existing in the node. The event identifier and its payload are required. This method returns the enumerated [CreateRequestResponse], being the identifier of a request when the event intends to update a governance, in which case it will be communicated to the rest of the nodes aware of the governance with the intention that they vote the change contained in the request. In case the event is from a conventional subject, the system returns the created event.

Possible errors

This method will return ApiError::EventCreationError if it has not been possible to create the event, while [ApiError::InternalError] will be obtained if an internal error has occurred during the management of the operation. If it has not been possible to create the signature accompanying the creation of the event with the node identity, then ApiError::SignError will be obtained.

fn get_all_subjects<'life0, 'async_trait>( &'life0 self, namespace: String, from: Option<String>, quantity: Option<i64> ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Allows to get all subjects that are known to the current node, regardless of their governance. Paging can be performed using the optional arguments from and quantity. Regarding the first one, note that it admits negative values, in which case the paging is performed in the opposite direction starting from the end of the collection. Note that this method also returns the subjects that model governance.

Possible errors

• [ApiError::InternalError] if an internal error occurred during the execution of the operation.

fn get_all_governances<'life0, 'async_trait>( &'life0 self, namespace: String, from: Option<String>, quantity: Option<i64> ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

It allows to obtain all the subjects that model existing governance in the node.

Possible errors

• [ApiError::InternalError] if an internal error occurred during the execution of the operation.

fn get_event_of_subject<'life0, 'async_trait>( &'life0 self, subject_id: DigestIdentifier, from: Option<i64>, quantity: Option<i64> ) -> Pin<Box<dyn Future<Output = Result<Vec<Event>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Allows to obtain events from a specific subject previously existing in the node. Paging can be performed by means of the optional arguments from and quantity. Regarding the former, it should be noted that negative values are allowed, in which case the paging is performed in the opposite direction starting from the end of the string.

Possible errors

• ApiError::InvalidParameters if the specified subject identifier does not match a valid DigestIdentifier.

fn create_subject<'life0, 'async_trait>( &'life0 self, governance_id: DigestIdentifier, schema_id: String, namespace: String ) -> Pin<Box<dyn Future<Output = Result<DigestIdentifier, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Allows to create a new subject in the node, being its owner the node in question.

Possible errors

• [ApiError::InternalError] if an internal error occurred during the execution of the operation.
• ApiError::SignError if it has not been possible to create the signature that accompanies the creation of the event with the identity of the node.
• ApiError::EventCreationError if it has not been possible to create the subject, for example, because its governance does not exist.
• ApiError::InvalidParameters if the specified governance identifier does not match a valid DigestIdentifier.

fn get_subject<'life0, 'async_trait>( &'life0 self, subject_id: DigestIdentifier ) -> Pin<Box<dyn Future<Output = Result<SubjectData, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Allows to obtain a specified subject by specifying its identifier.

Possible errors

• ApiError::InvalidParameters if the specified identifier does not match a valid DigestIdentifier.
• ApiError::NotFound if the subject does not exist.

fn create_governance<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<DigestIdentifier, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Method for creating governance in the system.

Possible errors

• [ApiError::InternalError] if an internal error occurred during the execution of the operation.
• ApiError::SignError if it has not been possible to create the signature that accompanies the creation of the event with the identity of the node.
• ApiError::EventCreationError if it has not been possible to create the governance.

fn shutdown<'async_trait>( self ) -> Pin<Box<dyn Future<Output = Result<(), ApiError>> + Send + 'async_trait>>where Self: 'async_trait,

Stops the node, consuming the instance on the fly. This implies that any previously created API or [NotificationHandler] instances will no longer be functional.

fn add_preauthorize_subject<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, subject_id: &'life1 DigestIdentifier, providers: &'life2 HashSet<KeyIdentifier> ) -> Pin<Box<dyn Future<Output = Result<(), ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

fn expecting_transfer<'life0, 'async_trait>( &'life0 self, subject_id: DigestIdentifier ) -> Pin<Box<dyn Future<Output = Result<KeyIdentifier, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn get_validation_proof<'life0, 'async_trait>( &'life0 self, subject_id: DigestIdentifier ) -> Pin<Box<dyn Future<Output = Result<HashSet<Signature>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn get_governance_subjects<'life0, 'async_trait>( &'life0 self, governance_id: DigestIdentifier, from: Option<String>, quantity: Option<i64> ) -> Pin<Box<dyn Future<Output = Result<Vec<SubjectData>, ApiError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,